+++ lib/plugins/aws/configCredentials.js
@@ -2,7 +2,7 @@
 
 const BbPromise = require('bluebird');
 const os = require('os');
-const credentials = require('../utils/credentials');
+const credentials = require('./utils/credentials');
 
 class AwsConfigCredentials {
   constructor(serverless, options) {
