+++ lib/plugins/index.js
@@ -20,7 +20,7 @@ module.exports = [
   require('./plugin/uninstall/uninstall.js'),
   require('./plugin/list/list.js'),
   require('./plugin/search/search.js'),
-  require('./aws/configCredentials/awsConfigCredentials.js'),
+  require('./aws/configCredentials.js'),
   require('./aws/provider/awsProvider.js'),
   require('./aws/common/index.js'),
   require('./aws/package/index.js'),
